Refactor project structure. Now uses automatic setup solution
