+++ CoreLib.sln
@@ -1,0 +1,16 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CoreLib", "CoreLib\CoreLib.csproj", "{A1C9C7BE-C448-4EA1-9793-7275B13162CF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A1C9C7BE-C448-4EA1-9793-7275B13162CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A1C9C7BE-C448-4EA1-9793-7275B13162CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A1C9C7BE-C448-4EA1-9793-7275B13162CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A1C9C7BE-C448-4EA1-9793-7275B13162CF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
