+++ CoreLib/Managers/GameManagers.cs
