+++ CoreLib/Managers/Localization.cs
