+++ CoreLib/Patches/ManagerPatch.cs
