+++ CoreLib/Patches/Rewired_Patch.cs
