+++ CoreLib/Util/RewiredExtensions.cs
