+++ README.md
@@ -1,4 +1,38 @@
 # CoreLib
-This library can be used to get some main objects of the game.
+A modding library for Core Keeper. Provides features that makes modding Core Keeper easier.
 
-WIP
+# List of features
+- Get main objects of the game, like `Managers`, `Players`
+- Add new localization terms
+- Add custom Rewired keybinds
+
+# Installation
+### With Mod Manager
+
+Simply open the mod manager (if you don't have it install it [here](https://dsp.thunderstore.io/package/ebkr/r2modman/)), select **CoreLib by CoreMods**, then **Download**.
+
+If prompted to download with dependencies, select `Yes`.
+Then just click **Start modded**, and the game will run with the mod installed.
+
+### Manually
+Install BepInEx Pack from [here](https://core-keeper.thunderstore.io/package/BepInEx/BepInExPack_Core_Keeper/)<br/>
+
+Unzip all files into `Core Keeper\BepInEx\plugins\CoreLib/` (Create folder named `CoreLib`)<br/>
+
+This library is still WIP.
+
+## Changelog
+<details>
+<summary>Changelog</summary>
+
+### v0.1.1
+- Now supports dedicated servers
+
+### v0.1.0
+- Added Localization helper class
+- Added RewiredKeybinds helper class
+- Improve README
+
+### v0.0.1
+- Initial Release
+</details>
