+++ Staging/manifest.json
@@ -1,7 +1,7 @@
 {
 	"name": "CoreLib",
 	"description": "A library that makes modding Core Keeper easier",
-	"version_number": "0.1.0",
+	"version_number": "0.1.1",
 	"dependencies": 
 	[
 		"BepInEx-BepInExPack_Core_Keeper-1.0.0"
